+++ qcio/__init__.py
@@ -7,21 +7,18 @@ __version__ = metadata.version(__name__)
 
 __all__ = [  # noqa: F405
     # Core Models
+    "Molecule",
     "Files",
-    "ComputedPropsBase",
     "OutputBase",
     "Files",
     "FileInput",
-    "FileResult",
-    "FileFailure",
+    "FileOutput",
     "Provenance",
     "Model",
     "CalcType",
     "ProgramInput",
     "SinglePointResults",
     "SinglePointOutput",
-    "SinglePointFailure",
-    "SinglePointComputedProps",
     "Wavefunction",
     "ProgramFailure",
     "ProgramArgs",