+++ tests/test_package_api.py
@@ -1,0 +1,6 @@
+from qcio import *  # noqa: F403
+
+
+def test_imports():
+    """Checking that import * above in module works."""
+    assert True